requirements.txt
anyio==3.5.0
asgiref==3.5.0
bidict==0.21.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
fastapi-mqtt==0.4.0
fastapi-socketio==0.0.9
gmqtt==0.6.11
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
orjson==3.6.7
pydantic==1.9.0
python-dotenv==0.19.2
python-engineio==4.3.1
python-multipart==0.0.5
python-socketio==5.5.2
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2
fastapi-mail
